setup unit tests
