+++ next.config.mjs
@@ -1,5 +1,5 @@
-/** @type {import("next").NextConfig} */
-module.exports = {
+/** @type {import('next').NextConfig} */
+const nextConfig = {
   images: {
     remotePatterns: [
       { hostname: "lh3.googleusercontent.com" },
@@ -8,3 +8,5 @@ module.exports = {
     ],
   },
 };
+
+export default nextConfig;
