+++ src/lib/crypto.ts
@@ -1,3 +1,4 @@
+import { getOwnedCoins } from "@/lib/utils";
 import { CoinMetadata } from "@/types";
 import type { Transaction } from "@prisma/client";
 
@@ -11,7 +12,7 @@ const fetchCrypto = async <T>(url: string): Promise<T> => {
       },
     }
   );
-  if (!res.ok) throw new Error("CoinMarketCap rate limited us");
+  if (!res.ok) throw new Error("Error fetching crypto data");
   const json = (await res.json()) as { data: T };
   return json.data;
 };
@@ -73,35 +74,10 @@ export const getMetadata = async (
   return json.data;
 };
 
-export const getOwnedCoins = async (transactions: Transaction[]) => {
+export const getWallet = async (transactions: Transaction[]) => {
   if (transactions.length === 0) return [];
 
-  const uniqueSymbols = [
-    ...new Set(transactions.map((transaction) => transaction.symbol)),
-  ];
-
-  const ownedCoins: { symbol: string; quantity: number }[] = [];
-
-  uniqueSymbols.forEach((symbol) => {
-    const symbolTransactions = transactions.filter(
-      (transaction) => transaction.symbol === symbol
-    );
-    const totalQuantity = symbolTransactions.reduce((total, transaction) => {
-      if (transaction.type === "BUY") {
-        return total + transaction.quantity;
-      }
-      return total - transaction.quantity;
-    }, 0);
-
-    if (totalQuantity > 0) {
-      ownedCoins.push({
-        symbol,
-        quantity: totalQuantity,
-      });
-    }
-  });
-
-  if (ownedCoins.length === 0) return [];
+  const ownedCoins = getOwnedCoins(transactions);
 
   const ownedCoinSymbols = ownedCoins.map((coin) => coin.symbol).join(",");
   const data = await fetchCrypto<{
@@ -126,8 +102,9 @@ export const getOwnedCoins = async (transactions: Transaction[]) => {
     const percentChange24h = data[symbol]?.quote.EUR.percent_change_24h;
     const percentChange7d = data[symbol]?.quote.EUR.percent_change_7d;
 
-    if (!currentPrice || !name || !percentChange24h)
+    if (!currentPrice || !name || !percentChange24h) {
       throw new Error(`Invalid symbol: ${symbol}`);
+    }
 
     return {
       symbol,
