+++ src/lib/utils.ts
@@ -1,3 +1,5 @@
+import { Transaction } from "@prisma/client";
+
 export const formatCurrency = (value: number) =>
   value.toLocaleString("fi-FI", {
     style: "currency",
@@ -18,3 +20,32 @@ export const getDeltaType = (percentChange: number) => {
 
 export const calculateCoinShare = (coinValue: number, portfolioValue: number) =>
   (coinValue / portfolioValue) * 100;
+
+export const getOwnedCoins = (transactions: Transaction[]) => {
+  const uniqueSymbols = [
+    ...new Set(transactions.map((transaction) => transaction.symbol)),
+  ];
+
+  const ownedCoins: { symbol: string; quantity: number }[] = [];
+
+  uniqueSymbols.forEach((symbol) => {
+    const symbolTransactions = transactions.filter(
+      (transaction) => transaction.symbol === symbol
+    );
+    const totalQuantity = symbolTransactions.reduce((total, transaction) => {
+      if (transaction.type === "BUY") {
+        return total + transaction.quantity;
+      }
+      return total - transaction.quantity;
+    }, 0);
+
+    if (totalQuantity > 0) {
+      ownedCoins.push({
+        symbol,
+        quantity: totalQuantity,
+      });
+    }
+  });
+
+  return ownedCoins;
+};
