+++ vitest.config.mjs
@@ -1,0 +1,9 @@
+/** @type {import('vite').UserConfig} */
+const config = {
+  plugins: [],
+  test: {
+    environment: "node",
+  },
+};
+
+export default config;
